requirements.txt
flask==2.1.2
irsdk==1.3.2
python-socketio==5.6.0